LayoutTests/inspector/styles/edit-inspector-stylesheet-expected.txt

Tests that adding a new rule creates inspector stylesheet resource and allows its live editing.

Text
Inspector stylesheet URL: inspector-stylesheet
Inspector stylesheet content:
#inspected {}

Setting new content
Inspector stylesheet content:
#inspected { background-color: green; }
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#inspected  { (inspector-stylesheet:1)
background-color: green;

[expanded] 
div  { (user agent stylesheet)
display: block;
